rename todo -> entity
